+++ CHANGES.rst
@@ -167,3 +167,4 @@ v1.1.9
 
 v1.1.10[unreleased]
 -----------
+* Implement Project.hidden_address field.